VOL-1858 Report link state for active UNI
Properly update UNI state and expose to core. This depends
on listening for new port state updates collected via OMCI
LAN LOS alarms.   Listening and acting only on these
allows only ports that are actually being used to be reported
to the core and onos.  Ultimately reducing the flow and
message count to only ports that are needed.

Note, unplugging and plugging back in the RG in the UNI
port will result in the removal and re-addition of flows.
This is more ideal but is a new aspect in testing and operations
that must be considered.

Also when core calls to get ofp port state do no just assume
LIVE, query and get the current operational state.

Additionally properly set the operational state.

Depends on https://gerrit.opencord.org/#/c/16981/

Also depends on bbsim being updated to send these LAN LOS
alarms in order to pass sanity tests

Start version 2.3.0-dev

Change-Id: I0d9e2b913347d74d46d52ab87eb501fc17378589