VOL-2171: Lock ont-g/ani only after all pptp/veip are locked
Also yield so the locking activity finishes before the
callbacks are called, so the MDS increment for the set
can be recorded in openomci

Ultimately this all prevents the MDS from getting out of sync
when disable is called and the locking is *halfway* done and
the omci state machine is prematurely stopped, causing some set responses
to be missed. which cases the MDS increment to be off.

This then later causes during an enable a forced reconcile when
we dont need it.  Having the MDS properly match speeds up re-enable.

Also remove custom get_mds_task thats not been used in quite some time.

Change-Id: Idfe295a61e8fc2810e8f0c956c5997c7a26b32ab