VOL-2524: Do not use omci link alarms for uni state
ONU sending omci link state alarms do not seem reliable enough
to drive dataplane provisioning.

The problem is two part:

1) omci alarms by default are unset (no alarms).  This includes
uni link state alarms.  This means by default all UNI are up.
Initially this means the core and BAL must provision flows for
ALL uni.

2) Given above the link state for unis is initially all up,
then all down (some staying down),
then one uni with an RG is up again. this messaging churn cause stress in other systems,
most notibly BAL on the olt.

There is a *need* for a down alarm to happen for an up alarm
to mean anything. And if these alarms are not reliably sent by the onu, we cannot depend
on them for provisioning.

Later (15 seconds) alarm reconciliation may pick up the alarm for uni that are permanently down
but if the uni that needs the down-then-up misses the down.. it never gets to act
on the "clear" or up alarm in order to provision.

So we assume all uni are up by default.  Also given there is no operator
requirement for multiple uni (and its load implications on the core), we only enable
the first uni.  Future features may decide how to know which uni to provision.

Also start version 2.3.2-dev

Change-Id: Idc1642bcd67d337c27ac59842d52dc6cc032eb08